feat(structure): add useDocumentPaneInspector hook
